requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
faster-whisper>=0.9.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.15
pydantic>=2.0.0
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx>=0.24.0  # Required for TestClient 